requirements.txt
# Core dependencies
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.3
numpy==1.25.2
sqlalchemy==2.0.23
# psycopg2-binary==2.9.9  # Installed via conda
yfinance==0.2.28
# tqdm==4.66.1  # Already installed via conda
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1
tenacity==8.2.3
aiohttp==3.9.1

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0 